Fixed an issue where the buttons were not corresponding to clicks
The original code used XML style to listen for a button click.  It does
not work and I did not bother investigating why but instead I
instantiated the buttons in my MainActivity and written my onClick
method there which I then call in the onCreate method.  Most common way
to add a listener to a button I believe.